Application goes back to the main App loop now
Before this commit our application would not offer the next
problem to solve. It gave you one and told you if the answer
was correct or wrong and didn't offer the next problem.

Now it gives you a problem to solve, tells you if the answer is correct
or wrong and after a few seconds it gives you the next problem.

To exit from the application, you need to close the window.

Added new bug in the process - correct answer is always on the third
button.