Added new colours to the app.
The correct score on top of the screen is now green and the wrong score is red.
When the user gets an answer correct they are presented with a red "Wrong!" or
a green "Correct".
At the end of 20 questions the user is presented by their total percentage
of correct answers which is red, if the user got below 50% or green, if the user got
above.